% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlserver.R
\name{db_insert_into.OdbcConnection}
\alias{db_insert_into.OdbcConnection}
\title{Todo: find a solution for storing rownames}
\usage{
\method{db_insert_into}{OdbcConnection}(con, table, values, ...)
}
\description{
Todo: find a solution for storing rownames
}
