% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-methods.R
\docType{methods}
\name{dbExecute}
\alias{dbExecute}
\alias{dbExecute,SQLServerConnection,character-method}
\alias{dbSendUpdate}
\title{Execute non-query SQL commands}
\usage{
dbExecute(conn, statement, ...)

\S4method{dbExecute}{SQLServerConnection,character}(conn, statement, ...)

dbSendUpdate(conn, statement, ...)
}
\arguments{
\item{conn}{a connection inheriting from \code{DBIConnection}}

\item{statement}{a string denoting the SQL command}
}
\value{
logical value representing whether or not the command has been
executed successfully.
}
\description{
Methods for this generic will execute non-query (i.e. non SELECT) SQL
commands like \code{CREATE TABLE} or \code{UPDATE}. The \code{dbSendUpdate}
method which was based on the same method in the \code{RJDBC} package will
be deprecated.
}
\examples{
\dontrun{
con <- dbConnect(RSQLServer::SQLServer(), "TEST")
dbExecute(con, "CREATE TABLE test (a integer)")
}
}

