% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DBDriver.R
\docType{methods}
\name{dbGetInfo,SQLServerDriver-method}
\alias{dbGetInfo,SQLServerDriver-method}
\title{Get driver info}
\usage{
\S4method{dbGetInfo}{SQLServerDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{Object created by \code{\link{SQLServer}}}

\item{...}{other arguments to methods. Not used here.}
}
\value{
A list containing the name and driver version used by \code{drv}
}
\description{
Get driver info
}
\examples{
\dontrun{
dbGetInfo(SQLServer())
}
}

