% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbListConnections,SQLServerDriver-method}
\alias{dbListConnections,SQLServerDriver-method}
\title{List active connections}
\usage{
\S4method{dbListConnections}{SQLServerDriver}(drv, ...)
}
\arguments{
\item{drv}{Object created by \code{\link{SQLServer}}}
}
\value{
An empty list as JDBC driver doesn't maintain a list of active
connections
}
\description{
List active connections
}
\examples{
\dontrun{
dbListConnection(SQLServer())
}
}

