% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_server_details}
\alias{get_server_details}
\title{Get server details from YAML file}
\usage{
get_server_details(server, file = NULL)
}
\arguments{
\item{server}{corresponds to the server name key in the YAML \code{file} and
should be a string.}

\item{file}{defaults to using \code{sql.yaml} in a user's \code{HOME}
directory (\code{Sys.getenv("HOME")}).}
}
\value{
a named list of \code{server} details if this is specified in the
\code{file}. It stops and returns an error if \code{port} and \code{type}
keys are not specified for found \code{server}. If the \code{file} does not
contain the \code{server} key an empty list is returned.
}
\description{
The \code{sql.yaml} file in a user's \code{HOME} directory can store
server details and login credentials (in plaintext). This works around
the instability associated with jTDS's single-sign on functionality.
The YAML file format is documented in this package's \code{README} file, while
an example is provided in \code{extdata/sql.yaml} (see example). At a
high level, each server should be documented in its own associative array
with each aspect of the server documented in an associative array.
}
\seealso{
\href{https://github.com/yaml/yaml}{YAML}
}
\keyword{internal}

