% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{have_test_server}
\alias{have_test_server}
\title{Checks availability of TEST server}
\usage{
have_test_server(type = "sqlserver")
}
\arguments{
\item{type}{specifies whether the server type is \code{"sqlserver"} (default)
or \code{"sybase"}}
}
\value{
boolean \code{TRUE} if TEST server details are available. Otherwise,
\code{FALSE}
}
\description{
SQL Server details can be specified in a \code{~/sql.yaml} file.
To be able to run examples and some tests in the package, it is necessary for
there to be a valid server with name TEST in this file.
}
\examples{
have_test_server()
}
\seealso{
\code{\link{get_server_details}} \code{\link{dbConnect,SQLServerDriver-method}}
}
