\name{rsqlserver-package}
\alias{rsqlserver-package}
\alias{rsqlserver}
\docType{package}
\title{
 SQL Server interface for R
}
\description{
Database Interface R driver for SQL server.  
            This package wraps the System.Data.SqlClient 
            assembly and provide a handy R interface.
}
\details{
\tabular{ll}{
Package: \tab rsqlserver\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-23\cr
License: \tab GPL-2\cr
}
}
\author{
Amine Gassem (agstudy) <contact@agstudy.com>
}

\references{
}

\keyword{ package }
\keyword{ SQL server }
\seealso{
\code{{RODBC}}
}
