% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{RSQLServer}
\alias{RSQLServer}
\alias{RSQLServer-package}
\title{RSQLServer}
\description{
Wrapper around RJDBC DBI applied specifically to SQL Server. You will need the
\href{https://www.java.com/en/download/index.jsp}{Java Runtime Environment (version 6 or higher)}.
}
\details{
If you intend to use integrated security (Windows Authentication) to
authenticate your server session, you will need to download \href{http://www.microsoft.com/en-us/download/details.aspx?id=11774}{Microsoft's JDBC Driver for SQL Server package}
and copy the bundled driver file (\code{sqljdbc_auth.dll}) to your java
library path (e.g. \code{C:\Program Files (x86)\Java\jre7\bin\client}).
}

