% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{setops}
\alias{intersect}
\alias{intersect.tbl_sqlserver}
\alias{setdiff}
\alias{setdiff.tbl_sqlserver}
\alias{setops}
\title{Intersect and setdiff methods}
\usage{
\method{intersect}{tbl_sqlserver}(x, y, copy = FALSE, ...)

\method{setdiff}{tbl_sqlserver}(x, y, copy = FALSE, ...)
}
\arguments{
\item{x, y}{objects to compare (ignoring order)}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{...}{other arguments passed on to methods}
}
\description{
Customised intersect and setdiff methods for the dplyr generics (which
override base package function definitions). These methods
only support SQL Server 2005 and greater.
}
\seealso{
\code{\link[dplyr]{setdiff}} \code{\link[dplyr]{intersect}}
}

