% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-methods.R
\docType{methods}
\name{SQLServer}
\alias{SQLServer}
\alias{dbConnect,SQLServerDriver-method}
\title{SQLServerDriver class and methods}
\usage{
SQLServer()

\S4method{dbConnect}{SQLServerDriver}(drv, alias = NULL, user = NULL,
  password = NULL, host = NULL, port = NULL, dbname = NULL, ...)
}
\arguments{
\item{drv}{An objected of class \code{\linkS4class{SQLServerDriver}}, or an
existing \code{\linkS4class{SQLServerConnection}}. If a connection, the
connection will be cloned.}

\item{alias}{the name of the server identified in the \code{~/sql.yaml} file. This
\code{alias} is described by a number of connection properties including the
\code{host} address, \code{portNumber}, \code{user}, \code{password} and \code{databaseName}.
Supported connection properties must have a corresponding setter in the
\href{https://docs.microsoft.com/en-us/sql/connect/jdbc/reference/sqlserverdatasource-members}{SQLServerDataSourceclass}.}

\item{user}{a string representing the user name}

\item{password}{a string representing the password used to login as the
\code{user}}

\item{host}{a string representing the address of the hosting server}

\item{port}{an integer representing the port that the \code{host} is listening on
(default: \code{1433L})}

\item{dbname}{a string representing the database to connect to on the \code{host}
(default: the default database on the \code{host})}

\item{...}{other supported connection properties}
}
\value{
\code{SQLServer()} returns an object of class
\code{SQLServerDriver}; \code{dbConnect()} returns a
\code{\linkS4class{SQLServerConnection}} object.
}
\description{
\code{SQLServer()} creates a \code{SQLServerDriver} object that is based on
on Microsoft's SQL Server JDBC driver while \code{dbConnect()} provides a
convenient interface to connecting to a SQL Server database using this
driver.
}
\examples{
\dontrun{
SQLServer()
}
# View sql.yaml file bundled in package
file <- system.file("extdata", "sql.yaml", package = "RSQLServer")
readLines(file)
# Connect using ~/sql.yaml file
\dontrun{
if (have_test_server()) {
 dbConnect(RSQLServer::SQLServer(), "TEST")
}
# Example where ~/sql.yaml does not exist or where the server
# is not in the YAML file.
dbConnect(RSQLServer::SQLServer(), server="11.1.111.11", port=1434,
   properties=list(useNTLMv2="true", domain="myco", user="me",
     password="asecret"))
}
}
\references{
\href{https://docs.microsoft.com/en-us/sql/connect/jdbc/reference/sqlserverdriver-class}{Microsoft SQLServerDriver class API doc}
}
