% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R
\docType{class}
\name{SQLServerConnection-class}
\alias{SQLServerConnection-class}
\alias{dbDataType,SQLServerConnection,ANY-method}
\alias{dbDisconnect,SQLServerConnection-method}
\alias{dbExistsTable,SQLServerConnection-method}
\alias{dbGetInfo,SQLServerConnection-method}
\alias{dbIsValid,SQLServerConnection-method}
\alias{dbListTables,SQLServerConnection-method}
\alias{dbSendQuery,SQLServerConnection,character-method}
\alias{dbWriteTable,SQLServerConnection-method}
\title{SQL Server Connection class}
\usage{
\S4method{dbGetInfo}{SQLServerConnection}(dbObj, ...)

\S4method{dbIsValid}{SQLServerConnection}(dbObj, ...)

\S4method{dbDisconnect}{SQLServerConnection}(conn, ...)

\S4method{dbSendQuery}{SQLServerConnection,character}(conn, statement, ...)

\S4method{dbDataType}{SQLServerConnection,ANY}(dbObj, obj, ...)

\S4method{dbWriteTable}{SQLServerConnection}(conn, name, value,
  overwrite = TRUE, append = FALSE)

\S4method{dbListTables}{SQLServerConnection}(conn, ...)

\S4method{dbExistsTable}{SQLServerConnection}(conn, name, ...)
}
\description{
SQL Server Connection class
}
\keyword{internal}

