% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBConnection.R
\docType{methods}
\name{dbConnect,SQLServerDriver-method}
\alias{dbConnect,SQLServerDriver-method}
\title{Connect to/disconnect from a SQL Server database.}
\usage{
\S4method{dbConnect}{SQLServerDriver}(drv, server, file = NULL,
  database = "", type = "sqlserver", port = "", properties = list())
}
\arguments{
\item{drv}{An objected of class \code{\linkS4class{SQLServerDriver}}, or an
existing \code{\linkS4class{SQLServerConnection}}. If a connection,
the connection will be cloned.}

\item{server}{the server address or recognised alias thereof.}

\item{file}{defaults to using the server details file in
\code{\$HOME/sql.yaml}. The server details including \code{type}, \code{port}
and any optional \code{properties} can be sourced from this file. If the
\code{server} name is found in \code{file}, the details therein are used
(and in particular, those provided in other arguments to this function are
ignored). The connection method prefers server details to be provided in
a \code{"sql.yaml"} file rather than provided as arguments to this function.
If you wish to specify the details as parameters, ensure that the \code{file}
argument is missing from the call.}

\item{database}{the name of the database hosted on the \code{server}. If an
empty string, a connection to the default database on \code{server} is
assumed.}

\item{type}{the server type. Must be either \code{"sqlserver"} or
\code{"sybase"}. Defaults to \code{"sqlserver"}.}

\item{port}{the TCP/IP default port. This will be coerced to a string. Defaults
to \code{1433} if an empty string.}

\item{properties}{One or more \href{http://jtds.sourceforge.net/faq.html}{optional connection properties.}
in a named list. Note if you intend to set the \code{useNTLMv2} property to \code{'true'}
from the default API value of \code{'false'}, you will need to make a specific
authentication driver available to the SQL Server driver, although this hasn't
worked particularly well in testing. See \code{\link{RSQLServer}} for more
details. Should you wish to use Windows authentication to connect to the server,
I recommend you set the following optional parameters: set \code{useNTLMv2} to
\code{'true'}, \code{domain} to your domain and \code{user} and
\code{password} to your username and password on \code{domain}. jTDS'
SSO functionality is flaky.}
}
\value{
a \code{linkS4Class{SQLServerConnection}}
}
\description{
Connect to/disconnect from a SQL Server database.
}
\examples{
# View sql.yaml file bundled in package
file <- system.file("extdata", "sql.yaml", package = "RSQLServer")
readLines(file)
# Connect using ~/sql.yaml file
if (have_test_server()) {
 dbConnect(SQLServer(), "TEST")
}
# Don't use file argument:
\dontrun{
dbConnect(SQLServer(), server="11.1.111.11", port=1434,
   properties=list(useNTLMv2="true", domain="myco", user="me",
     password="asecret"))
}
}

