% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBConnection.R
\docType{methods}
\name{dbConnect,SQLServerDriver-method}
\alias{dbConnect,SQLServerDriver-method}
\title{Connect to/disconnect from a SQL Server database.}
\usage{
\S4method{dbConnect}{SQLServerDriver}(drv, server, ...)
}
\arguments{
\item{drv}{An objected generated by \code{\link{SQLServer}}, or an existing
\code{\linkS4class{SQLServerConnection}}. If a connection, the connection
will be cloned.}

\item{server}{The address of the server to connect to.}

\item{...}{One or more \href{http://jtds.sourceforge.net/faq.html}{optional connection properties.}.
Note if you intend to set the \code{useNTLMv2} property to \code{'true'}
from the default API value of \code{'false'}, you will need to make a specific
authentication driver available to the SQL Server driver. See
\code{\link{RSQLServer}} for more details}
}
\value{
a \code{linkS4Class{SQLServerConnection}} object
}
\description{
Connect to/disconnect from a SQL Server database.
}
\examples{
\dontrun{
dbConnect(SQLServer(), 'ServerName')
}
}

