% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DBConnection.R
\docType{methods}
\name{dbIsValid,SQLServerConnection-method}
\alias{dbIsValid,SQLServerConnection-method}
\title{Checks whether Connection is closed}
\usage{
\S4method{dbIsValid}{SQLServerConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from \code{\linkS4class{SQLServerConnection}}.}

\item{...}{other parameters. Not used.}
}
\value{
logical \code{TRUE} if the connection is closed and vice-versa
}
\description{
Checks whether Connection is closed
}

