% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBResult.R
\docType{methods}
\name{dbIsValid,SQLServerResult-method}
\alias{dbIsValid,SQLServerResult-method}
\title{Checks whether ResultSet is closed}
\usage{
\S4method{dbIsValid}{SQLServerResult}(dbObj)
}
\arguments{
\item{dbObj}{An object inheriting from \code{\linkS4class{SQLServerResult}}.}
}
\value{
logical \code{TRUE} if the result set is closed and vice-versa
}
\description{
Checks whether ResultSet is closed
}

