% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbSendQuery,SQLServerConnection,character-method}
\alias{dbSendQuery,SQLServerConnection,character-method}
\title{Send query to SQL Server}
\usage{
\S4method{dbSendQuery}{SQLServerConnection,character}(conn, statement, ...,
  list = NULL)
}
\arguments{
\item{conn}{connection object}

\item{statement}{SQL statement to execute}

\item{...}{additional arguments to prepared statement substituted for "?"}

\item{list}{undocumented}
}
\value{
a \code{\linkS4class{SQLServerResult}} object
}
\description{
This is basically a copy of RJDBC's \code{\link[RJDBC:JDBCConnection-methods]{dbSendQuery}}
method for JDBCConnection.
}

