% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSQLServer.R
\docType{package}
\name{RSQLServer}
\alias{RSQLServer}
\alias{RSQLServer-package}
\title{RSQLServer}
\description{
Implements an R Database Interface (DBI) for SQL Server using Microsoft's
JDBC driver. While the necessary JDBC drivers are bundled with the package,
you will need to have at least version 8 of the \href{https://www.java.com/en/download/index.jsp}{Java RuntimeEnvironment} installed. Only SQL
Server 2008 and later are supported by the bundled driver.
}
\references{
\href{https://docs.microsoft.com/en-us/sql/connect/jdbc/system-requirements-for-the-jdbc-driver}{System requirements for Microsoft JDBC drivers}
}
