% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-methods.R
\docType{methods}
\name{SQLServer}
\alias{SQLServer}
\alias{dbConnect,SQLServerDriver-method}
\title{SQLServerDriver class and methods}
\usage{
SQLServer()

\S4method{dbConnect}{SQLServerDriver}(drv, server, file = NULL,
  database = NULL, type = NULL, port = NULL, properties = NULL)
}
\arguments{
\item{drv}{An objected of class \code{\linkS4class{SQLServerDriver}}, or an
existing \code{\linkS4class{SQLServerConnection}}. If a connection, the
connection will be cloned.}

\item{server}{the server address or recognised alias thereof.}

\item{file}{defaults to using the server details file in
\code{$HOME/sql.yaml}. The server details including \code{type},
\code{port} and any optional \code{properties} can be sourced from this
file. If the \code{server} name is found in \code{file}, the details
therein are used (and in particular, those provided in other arguments to
this function are ignored). The connection method prefers server details
to be provided in a \code{"sql.yaml"} file rather than provided as
arguments to this function. If you wish to specify the details as
parameters, ensure that either the \code{file} does not exist or that the
\code{server} details are not in the YAML file.}

\item{database}{the name of the database hosted on the \code{server}. If an
empty string or \code{NULL} (default), a connection to the default
database on \code{server} is assumed.}

\item{type}{the server type. Must be either \code{"sqlserver"} or
\code{"sybase"}. Defaults to \code{"sqlserver"} when set to \code{NULL}.}

\item{port}{the TCP/IP default port. This will be coerced to a string.
Defaults to \code{1433} if set to \code{NULL} or an empty string.}

\item{properties}{One or more [optional connection
properties](https://docs.microsoft.com/en-us/sql/connect/jdbc/setting-the-connection-properties).
in a named list (defaults to empty list).}
}
\value{
\code{SQLServer()} returns an object of class
\code{SQLServerDriver}; \code{dbConnect()} returns a
\code{\linkS4class{SQLServerConnection}} object.
}
\description{
\code{SQLServer()} creates a \code{SQLServerDriver} object that is based on
on Microsoft's SQL Server JDBC driver while \code{dbConnect()} provides a
convenient interface to connecting to a SQL Server database using this
driver.
}
\examples{
\dontrun{
SQLServer()
}
# View sql.yaml file bundled in package
file <- system.file("extdata", "sql.yaml", package = "RSQLServer")
readLines(file)
# Connect using ~/sql.yaml file
\dontrun{
if (have_test_server()) {
 dbConnect(RSQLServer::SQLServer(), "TEST")
}
# Example where ~/sql.yaml does not exist or where the server
# is not in the YAML file.
dbConnect(RSQLServer::SQLServer(), server="11.1.111.11", port=1434,
   properties=list(useNTLMv2="true", domain="myco", user="me",
     password="asecret"))
}
}
\references{
\href{https://docs.microsoft.com/en-us/sql/connect/jdbc/reference/sqlserverdriver-class}{Microsoft SQLServerDriver class API doc}
}
