% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBConnection.R
\docType{class}
\name{SQLServerConnection-class}
\alias{SQLServerConnection-class}
\alias{SQLServerConnection-method}
\alias{dbGetInfo,SQLServerConnection-method}
\alias{dbIsValid,SQLServerConnection-method}
\alias{dbSendQuery,}
\alias{dbSendQuery,SQLServerConnection,character-method}
\title{An S4 class to represent a SQL Server connection}
\usage{
\S4method{dbGetInfo}{SQLServerConnection}(dbObj, ...)

\S4method{dbIsValid}{SQLServerConnection}(dbObj, ...)

\S4method{dbSendQuery}{SQLServerConnection,character}(conn, statement, ...,
  list = NULL)
}
\arguments{
\item{dbObj}{a \code{\linkS4class{SQLServerConnection}}}

\item{...}{additional arguments to prepared statement substituted for "?"}

\item{conn}{a \code{\linkS4class{SQLServerConnection}}}

\item{statement}{SQL statement to execute}

\item{list}{undocumented}
}
\value{
a \code{\linkS4class{SQLServerResult}} object
}
\description{
This class extends the \code{\link[RJDBC:JDBCConnection-class]{JDBCConnection}}
class to represent a SQL Server connection.

This is basically a copy of RJDBC's \code{\link[RJDBC:JDBCConnection-methods]{dbSendQuery}}
method for JDBCConnection.
}
\section{Slots}{

\describe{
\item{\code{jc}}{Java object representing the connection.}

\item{\code{identifier.quote}}{quote character for a SQL Server identifier can be a
single quotation mark (\code{\'}), a left or right bracket (\code{[]}), or a
double quotation mark (\code{\"}). Usually inherited from
\code{\linkS4class{SQLServerDriver}}.}
}}

