% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SQLServerConnection-class}
\alias{SQLServerConnection-class}
\title{An S4 class to represent a SQL Server connection}
\description{
This class extends the \code{\link[RJDBC:JDBCConnection-class]{JDBCConnection}}
class to represent a SQL Server connection.
}
\section{Slots}{

\describe{
\item{\code{jc}}{Java object representing the connection.}

\item{\code{identifier.quote}}{quote character for a SQL Server identifier can be a
single quotation mark (\code{\'}), a left or right bracket (\code{[]}), or a
double quotation mark (\code{\"}). Usually inherited from
\code{\linkS4class{SQLServerDriver}}.}
}}

