% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R
\docType{class}
\name{SQLServerConnection-class}
\alias{SQLServerConnection-class}
\alias{dbDataType,SQLServerConnection-method}
\alias{dbGetInfo,SQLServerConnection-method}
\alias{dbGetQuery,SQLServerConnection,character-method}
\alias{dbIsValid,SQLServerConnection-method}
\alias{dbSendQuery,SQLServerConnection,character-method}
\alias{dbSendUpdate,SQLServerConnection,character-method}
\alias{show,SQLServerConnection-method}
\title{An S4 class to represent a SQL Server connection}
\usage{
\S4method{dbGetInfo}{SQLServerConnection}(dbObj, ...)

\S4method{show}{SQLServerConnection}(object)

\S4method{dbIsValid}{SQLServerConnection}(dbObj, ...)

\S4method{dbSendQuery}{SQLServerConnection,character}(conn, statement, ...,
  list = NULL)

\S4method{dbGetQuery}{SQLServerConnection,character}(conn, statement, ...)

\S4method{dbDataType}{SQLServerConnection}(dbObj, obj, ...)

\S4method{dbSendUpdate}{SQLServerConnection,character}(conn, statement, ...,
  list = NULL)
}
\arguments{
\item{...}{additional arguments to prepared statement substituted for "?"}

\item{object}{a \code{\linkS4class{SQLServerConnection}} object}

\item{statement}{SQL statement to execute}

\item{list}{undocumented}

\item{obj}{An R object whose SQL type we want to determine}
}
\value{
a \code{\linkS4class{SQLServerResult}} object
}
\description{
Extends the \code{\link[RJDBC:JDBCConnection-class]{JDBCConnection}}
class.

This is basically a copy of RJDBC's \code{\link[RJDBC:JDBCConnection-methods]{dbSendQuery}}
method for JDBCConnection, except that this returns a
\code{\linkS4class{SQLServerResult}} rather than a JDBCResult.
}
\keyword{internal}

