% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R, R/dbi-tables.R
\docType{class}
\name{SQLServerConnection-class}
\alias{SQLServerConnection-class}
\alias{dbGetInfo,SQLServerConnection-method}
\alias{dbIsValid,SQLServerConnection-method}
\alias{dbDisconnect,SQLServerConnection-method}
\alias{dbSendQuery,SQLServerConnection,character-method}
\alias{dbSendStatement,SQLServerConnection,character-method}
\alias{dbSendUpdate}
\alias{dbReadTable,SQLServerConnection,character-method}
\alias{dbDataType,SQLServerConnection-method}
\alias{dbListTables,SQLServerConnection-method}
\alias{dbExistsTable,SQLServerConnection-method}
\alias{dbRemoveTable,SQLServerConnection-method}
\alias{dbBegin,SQLServerConnection-method}
\alias{dbCommit,SQLServerConnection-method}
\alias{dbRollback,SQLServerConnection-method}
\alias{dbWriteTable,SQLServerConnection-method}
\alias{sqlCreateTable,SQLServerConnection-method}
\title{SQL Server Connection class}
\usage{
\S4method{dbGetInfo}{SQLServerConnection}(dbObj, ...)

\S4method{dbIsValid}{SQLServerConnection}(dbObj, ...)

\S4method{dbDisconnect}{SQLServerConnection}(conn, ...)

\S4method{dbSendQuery}{SQLServerConnection,character}(conn, statement,
  params = NULL, ..., batch = FALSE)

\S4method{dbSendStatement}{SQLServerConnection,character}(conn, statement,
  params = NULL, ..., batch = FALSE)

dbSendUpdate(conn, statement, ...)

\S4method{dbReadTable}{SQLServerConnection,character}(conn, name, ...)

\S4method{dbDataType}{SQLServerConnection}(dbObj, obj, ...)

\S4method{dbListTables}{SQLServerConnection}(conn, pattern = "\%", ...)

\S4method{dbExistsTable}{SQLServerConnection}(conn, name, ...)

\S4method{dbRemoveTable}{SQLServerConnection}(conn, name, ...)

\S4method{dbBegin}{SQLServerConnection}(conn, ...)

\S4method{dbCommit}{SQLServerConnection}(conn, ...)

\S4method{dbRollback}{SQLServerConnection}(conn, ...)

\S4method{dbWriteTable}{SQLServerConnection}(conn, name, value,
  row.names = NA, overwrite = FALSE, append = FALSE, field.types = NULL,
  temporary = FALSE, batch = FALSE)

\S4method{sqlCreateTable}{SQLServerConnection}(con, table, fields,
  row.names = NA, temporary = FALSE, ...)
}
\arguments{
\item{batch}{logical, indicates whether uploads (e.g., 'INSERT' or
'UPDATE') should be uploaded in batches, potentially much faster
than by individual rows (the default).}
}
\description{
SQL Server Connection class
}
\keyword{internal}
