% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R
\docType{class}
\name{SQLServerDriver-class}
\alias{SQLServerDriver-class}
\alias{dbGetInfo,SQLServerDriver-method}
\title{An S4 class to represent a SQL Server driver}
\usage{
\S4method{dbGetInfo}{SQLServerDriver}(dbObj, ...)
}
\description{
Extends the \code{\link[RJDBC:JDBCDriver-class]{JDBCDriver}} class and
uses jTDS driver.
}
\seealso{
\code{\link{SQLServer}}
}
\keyword{internal}

