% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R
\docType{class}
\name{SQLServerDriver-class}
\alias{SQLServerDriver-class}
\alias{dbGetInfo,SQLServerDriver-method}
\alias{dbIsValid,SQLServerDriver-method}
\alias{dbUnloadDriver,SQLServerDriver-method}
\title{SQL Server Driver class}
\usage{
\S4method{dbGetInfo}{SQLServerDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SQLServerDriver}(drv, ...)

\S4method{dbIsValid}{SQLServerDriver}(dbObj, ...)
}
\description{
SQL Server Driver class
}
\seealso{
\code{\link{SQLServer}}
}
\keyword{internal}

