% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBResult.R
\docType{class}
\name{SQLServerResult-class}
\alias{SQLServerResult-class}
\alias{dbIsValid,SQLServerResult-method}
\title{An S4 class to represent a SQL Server result set}
\usage{
\S4method{dbIsValid}{SQLServerResult}(dbObj)
}
\arguments{
\item{dbObj}{a \code{\linkS4class{SQLServerResult}} object}
}
\description{
This class extends the \code{\link[RJDBC:JDBCResult-class]{JDBCResult}}
class to represent a SQL Server result set
}
\section{Slots}{

\describe{
\item{\code{jr}}{Java reference to the JDBC result set}

\item{\code{md}}{Java reference to the JDBC result set meta data}

\item{\code{pull}}{Java reference to the JDBC result pull helper class (can be null
reference before first pull)}

\item{\code{stat}}{Java reference to the JDBC statement which generated this result}
}}

