% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-classes.R, R/dbi-methods.R
\docType{class}
\name{SQLServerResult-class}
\alias{SQLServerResult-class}
\alias{dbClearResult,SQLServerResult-method}
\alias{dbColumnInfo,SQLServerResult-method}
\alias{dbFetch,SQLServerResult,numeric-method}
\alias{dbGetInfo,SQLServerResult-method}
\alias{dbHasCompleted,SQLServerResult-method}
\alias{dbIsValid,SQLServerResult-method}
\alias{fetch,SQLServerResult,numeric-method}
\title{SQL Server Result class}
\usage{
\S4method{dbIsValid}{SQLServerResult}(dbObj)

\S4method{dbFetch}{SQLServerResult,numeric}(res, n = -1, ...)

\S4method{fetch}{SQLServerResult,numeric}(res, n = -1, block = 2048L, ...)

\S4method{dbGetInfo}{SQLServerResult}(dbObj, ...)

\S4method{dbColumnInfo}{SQLServerResult}(res, ...)

\S4method{dbHasCompleted}{SQLServerResult}(res, ...)

\S4method{dbClearResult}{SQLServerResult}(res, ...)
}
\description{
SQL Server Result class
}
\keyword{internal}

