% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SQLServerResult-class}
\alias{SQLServerResult-class}
\title{An S4 class to represent a SQL Server result set}
\description{
This class extends the \code{\link[RJDBC:JDBCResult-class]{JDBCResult}}
class to represent a SQL Server result set
}
\section{Slots}{

\describe{
\item{\code{jr}}{Java reference to the JDBC result set}

\item{\code{md}}{Java reference to the JDBC result set meta data}

\item{\code{pull}}{Java reference to the JDBC result pull helper class (can be null
reference before first pull)}

\item{\code{stat}}{Java reference to the JDBC statement which generated this result}
}}

